backend/requirements.txt
Flask
Werkzeug
Jinja2
numpy
opencv-contrib-python
pytesseract
Pillow
deep-translator
pandas
streamlit
python-docx
python-pptx
fpdf
csv